requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.5
django-bootstrap4==23.2
django-ckeditor==6.7.0
django-js-asset==2.1.0
Pillow==10.0.1
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
